Update operating_system.rb to avoid early load of etc.so
Fixes #388
Fixes https://github.com/ruby/etc/issues/29

+++ resources/files/operating_system.rb
@@ -54,8 +54,8 @@ Gem.pre_install do |gem_installer|
 end
 
 begin
-  config_fname = Gem::ConfigFile::SYSTEM_WIDE_CONFIG_FILE
-  unless File.exist?(config_fname)
+  config_dir = ENV['ProgramData']
+  if config_dir && !File.exist?(config_fname = File.join(config_dir, "gemrc"))
     File.open(config_fname, File::CREAT | File::EXCL | File::WRONLY) do |fd|
       fd.write <<-EOT
 # This is the system wide config file for Rubygems.
